client/python/projectairsim/src/projectairsim/autonomy/serving/requirements.txt
onnxruntime-gpu
pydantic
python-multipart
fastapi
uvicorn
